+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.6] - 2016-02-26
+### Added
+- Added unit tests for `Helper` class.
+
+### Fixed
+- Renamed `Helper::parse_params` to camelCase `Helper::parseParams`.
+
 ## [0.1.5] - 2016-02-26
 ### Added
 - Added `Helper` class that provides a static method to parse parameters.
@@ -29,6 +36,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release to GitHub.
 
+[0.1.6]: https://github.com/brightnucleus/invoker/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/brightnucleus/invoker/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/brightnucleus/invoker/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/brightnucleus/invoker/compare/v0.1.2...v0.1.3